fix: fail closed external move mutations
