+++ docs/adr/README.md
@@ -5,9 +5,9 @@ setup or operations guides, not in this index.
 
 | ADR                                                                                                                | Current status                                                                         | Relationship                                                                |
 | ------------------------------------------------------------------------------------------------------------------ | -------------------------------------------------------------------------------------- | --------------------------------------------------------------------------- |
-| [External document roots](ADR-External-Document-Roots.md)                                                          | Accepted and implemented; handoff and local move contracts later amended               | Base authorization, confinement and read/handoff contract                   |
+| [External document roots](ADR-External-Document-Roots.md)                                                          | Accepted and implemented; handoff plus diagnostic local-move planning later amended    | Base authorization, confinement and read/handoff contract                   |
 | [Governed HTTP delivery](ADR-HTTP-External-Artifact-Delivery.md)                                                   | Accepted and implemented on `main` for authenticated loopback; remote remains pilot    | Adds `http_ticket`; HTTP mutation remains denied                            |
-| [External reference integrity](ADR-External-Reference-Integrity.md) / [FR](ADR-External-Reference-Integrity.fr.md) | Accepted and implemented for a local stdio pilot                                       | Same-root file move, exact ÉLYSIA reference repair, journal and rollback    |
+| [External reference integrity](ADR-External-Reference-Integrity.md) / [FR](ADR-External-Reference-Integrity.fr.md) | Accepted; diagnostic local-stdio surfaces implemented, mutation deferred               | Same-root diagnostic plan/status, redacted evidence and deferred mutation   |
 | [Operon Bridge](ADR-Operon-Bridge.md)                                                                              | Accepted for bounded pilot                                                             | Versioned task-domain bridge and guarded mutation contract                  |
 | [Common governed operation runtime](ADR-Common-Operation-Runtime.md)                                               | Accepted and released in 2.6.0; public atomic-note projection and live canary complete | Shared plan, apply, status, receipt and exact-plan recovery vocabulary      |
 | [P1 governed frontmatter projection](ADR-Governed-Frontmatter-P1.md)                                               | Accepted, implemented and released in 2.7.0                                            | Source-preserving domain compiler projected over the released P0 runtime    |
