+++ package.json
@@ -144,7 +144,7 @@
     "test:tool-routing": "npm run build && node scripts/test-tool-routing-contract.mjs && node scripts/smoke-headless-readonly.mjs",
     "test:log-privacy": "npm run build && node scripts/test-log-privacy.mjs && node scripts/test-format-validation-privacy.mjs && node scripts/test-mcp-sdk-tool-error-boundary.mjs && node scripts/test-batch-item-error-privacy.mjs && node scripts/test-bases-upsert-row-binding.mjs && node scripts/test-bases-legacy-failure-contract.mjs && node scripts/test-runtime-status-public.mjs && node scripts/test-http-configuration-privacy.mjs",
     "test:local-rest-api-compat": "node scripts/test-local-rest-api-compat.mjs",
-    "test:external-roots": "npm run build && node scripts/test-external-reference-parser.mjs && node scripts/test-external-move-integrity.mjs && node scripts/test-external-move-journal-concurrency.mjs && node scripts/test-backend-vault-adapter-cas.mjs && node scripts/test-obsidian-cas.mjs && node scripts/test-obsidian-search-replace-cas.mjs && node scripts/test-external-roots.mjs && node scripts/test-stdio-proxy-external-roots.mjs && node scripts/test-http-external-handoff.mjs",
+    "test:external-roots": "npm run build && node scripts/test-external-reference-parser.mjs && node scripts/test-external-move-integrity.mjs && node scripts/test-external-move-status-security.mjs && node scripts/test-external-move-adapter-projections.mjs && node scripts/test-external-move-journal-concurrency.mjs && node scripts/test-backend-vault-adapter-cas.mjs && node scripts/test-obsidian-cas.mjs && node scripts/test-obsidian-search-replace-cas.mjs && node scripts/test-external-roots.mjs && node scripts/test-stdio-proxy-external-roots.mjs && node scripts/test-http-external-handoff.mjs",
     "test:external-roots:proxy": "npm run build && node scripts/test-stdio-proxy-external-roots.mjs",
     "test:http-external-handoff": "npm run build && node scripts/test-http-external-handoff.mjs",
     "test:http-multiclient": "npm run build && node scripts/test-http-multiclient-rate-limiting.mjs && node scripts/test-http-session-bounds.mjs",
@@ -164,8 +164,8 @@
     "test:launchers": "npm run build && node scripts/test-launchers.mjs",
     "test:external-reference-parser": "node scripts/test-external-reference-parser.mjs",
     "test:obsidian-cas": "node scripts/test-obsidian-cas.mjs && node scripts/test-obsidian-search-replace-cas.mjs",
-    "test:external-move": "node scripts/test-external-move-integrity.mjs",
-    "test:operation-runtime": "npm run build && node scripts/test-external-move-integrity.mjs && node scripts/test-obsidian-note-replace-operation.mjs && node scripts/test-base-config-p2-compiler.mjs && node scripts/test-governed-base-formula-operation.mjs && node scripts/test-canvas-p3-compiler.mjs && node scripts/test-governed-canvas-operation.mjs",
+    "test:external-move": "npm run build && node scripts/test-external-move-integrity.mjs && node scripts/test-external-move-status-security.mjs && node scripts/test-external-move-adapter-projections.mjs",
+    "test:operation-runtime": "npm run build && node scripts/test-external-move-integrity.mjs && node scripts/test-external-move-status-security.mjs && node scripts/test-external-move-adapter-projections.mjs && node scripts/test-obsidian-note-replace-operation.mjs && node scripts/test-base-config-p2-compiler.mjs && node scripts/test-governed-base-formula-operation.mjs && node scripts/test-canvas-p3-compiler.mjs && node scripts/test-governed-canvas-operation.mjs",
     "test:governed-note-replace-mcp": "npm run build && node scripts/test-governed-note-replace-mcp.mjs",
     "test:governed-note-replace-http": "npm run build && node scripts/test-governed-note-replace-http.mjs",
     "smoke:atomic-note-mcp-live": "npm run build && node scripts/smoke-atomic-note-mcp-live.mjs",
