+++ src/services/externalRootsService.ts
@@ -2,7 +2,6 @@ import { createHash, randomUUID } from "node:crypto";
 import { rmSync } from "node:fs";
 import {
   type FileHandle,
-  link,
   lstat,
   mkdtemp,
   open,
@@ -12,7 +11,6 @@ import {
   rename,
   rm,
   stat,
-  unlink,
   writeFile,
 } from "node:fs/promises";
 import os from "node:os";
@@ -152,6 +150,26 @@ export type ExternalEntry = {
   modifiedAt?: string;
 };
 
+export const EXTERNAL_MOVE_MUTATION_UNAVAILABLE_REASON =
+  "native_handle_relative_mutation_unavailable" as const;
+
+export function moveMutationStatus(): {
+  mutationAvailable: false;
+  mutationUnavailableReason: typeof EXTERNAL_MOVE_MUTATION_UNAVAILABLE_REASON;
+} {
+  return {
+    mutationAvailable: false,
+    mutationUnavailableReason: EXTERNAL_MOVE_MUTATION_UNAVAILABLE_REASON,
+  };
+}
+
+export function externalMoveMutationUnavailableError(): ExternalRootError {
+  return new ExternalRootError(
+    "unsupported",
+    `External move mutation is unavailable: ${EXTERNAL_MOVE_MUTATION_UNAVAILABLE_REASON}.`,
+  );
+}
+
 export type ExternalMoveSnapshot = {
   rootId: string;
   sourceRelativePath: string;
@@ -160,13 +178,6 @@ export type ExternalMoveSnapshot = {
   modifiedAt: string;
   sha256: string;
 };
-
-type ExternalMoveIdentity = {
-  dev: bigint;
-  ino: bigint;
-};
-
-type SafeMovePathState = "absent" | "matching" | "mismatch";
 
 const textExtensions = new Set([
   ".txt",
@@ -604,240 +615,18 @@ export class ExternalRootsService {
   }
 
   async applyMove(snapshot: ExternalMoveSnapshot): Promise<void> {
-    const current = await this.planMove(
-      snapshot.rootId,
-      snapshot.sourceRelativePath,
-      snapshot.targetRelativePath,
-    );
-    if (
-      current.size !== snapshot.size ||
-      current.modifiedAt !== snapshot.modifiedAt ||
-      current.sha256 !== snapshot.sha256
-    ) {
-      throw new ExternalRootError(
-        "precondition_failed",
-        "The external source changed after the move was planned.",
-      );
-    }
-    const runtime = this.requireCapability(snapshot.rootId, "move");
-    const sourcePath = await this.resolvePath(
-      runtime,
-      snapshot.sourceRelativePath,
-    );
-    const targetPath = await this.resolveNewTarget(
-      runtime,
-      snapshot.targetRelativePath,
-    );
-    let linkedIdentity: ExternalMoveIdentity | undefined;
-    try {
-      // link() is no-clobber and keeps the operation on one filesystem. The
-      // source is removed only after the target identity is verified.
-      await link(sourcePath, targetPath);
-      const sourceStat = await lstat(sourcePath, { bigint: true });
-      const targetStat = await lstat(targetPath, { bigint: true });
-      linkedIdentity = { dev: targetStat.dev, ino: targetStat.ino };
-      const rootPath = await this.canonicalRoot(runtime);
-      const canonicalTarget = await realpath(targetPath);
-      if (
-        !sourceStat.isFile() ||
-        !targetStat.isFile() ||
-        sourceStat.isSymbolicLink() ||
-        targetStat.isSymbolicLink() ||
-        sourceStat.dev !== targetStat.dev ||
-        sourceStat.ino !== targetStat.ino ||
-        !isInsideRoot(rootPath, canonicalTarget)
-      ) {
-        throw new ExternalRootError(
-          "non_verifiable",
-          "The external move target is not the verified source object.",
-        );
-      }
-      const linkedState = await this.safeMovePathState(
-        runtime,
-        snapshot.targetRelativePath,
-        snapshot,
-      );
-      if (linkedState !== "matching") {
-        throw new ExternalRootError(
-          "precondition_failed",
-          "The linked external move target does not match the planned source.",
-        );
-      }
-      await this.unlinkIfIdentity(
-        runtime,
-        snapshot.sourceRelativePath,
-        linkedIdentity,
-      );
-    } catch (error) {
-      const code =
-        typeof error === "object" && error && "code" in error
-          ? String(error.code)
-          : "";
-      if (code === "EEXIST") {
-        throw new ExternalRootError(
-          "target_exists",
-          "The external move target already exists.",
-        );
-      }
-      if (linkedIdentity) {
-        try {
-          await this.unlinkIfIdentity(
-            runtime,
-            snapshot.targetRelativePath,
-            linkedIdentity,
-          );
-        } catch {
-          throw new ExternalRootError(
-            "non_verifiable",
-            "The external move failed and its linked target could not be cleaned safely.",
-          );
-        }
-      }
-      if (error instanceof ExternalRootError) throw error;
-      throw new ExternalRootError(
-        code === "EXDEV" || code === "ENOTSUP"
-          ? "unsupported"
-          : "non_verifiable",
-        "The filesystem cannot complete a verified no-clobber move.",
-      );
-    }
+    void snapshot;
+    throw externalMoveMutationUnavailableError();
   }
 
   async rollbackMove(snapshot: ExternalMoveSnapshot): Promise<void> {
-    const runtime = this.requireCapability(snapshot.rootId, "move");
-    const sourceRelativePath = normalizeRelativePath(
-      snapshot.sourceRelativePath,
-    );
-    const targetRelativePath = normalizeRelativePath(
-      snapshot.targetRelativePath,
-    );
-    const sourceState = await this.safeMovePathState(
-      runtime,
-      sourceRelativePath,
-      snapshot,
-    );
-    if (sourceState !== "absent") {
-      throw new ExternalRootError(
-        "target_exists",
-        "The original external source path is already occupied.",
-      );
-    }
-    const targetState = await this.safeMovePathState(
-      runtime,
-      targetRelativePath,
-      snapshot,
-    );
-    if (targetState !== "matching") {
-      throw new ExternalRootError(
-        targetState === "absent" ? "not_found" : "precondition_failed",
-        "The moved external file is absent or changed and cannot be rolled back safely.",
-      );
-    }
-    const sourcePath = await this.resolveNewTarget(runtime, sourceRelativePath);
-    const targetPath = await this.resolvePath(runtime, targetRelativePath);
-    let restoredIdentity: ExternalMoveIdentity | undefined;
-    try {
-      await link(targetPath, sourcePath);
-      const targetStat = await lstat(targetPath, { bigint: true });
-      const sourceStat = await lstat(sourcePath, { bigint: true });
-      restoredIdentity = { dev: sourceStat.dev, ino: sourceStat.ino };
-      const rootPath = await this.canonicalRoot(runtime);
-      const canonicalSource = await realpath(sourcePath);
-      if (
-        !targetStat.isFile() ||
-        !sourceStat.isFile() ||
-        targetStat.isSymbolicLink() ||
-        sourceStat.isSymbolicLink() ||
-        targetStat.dev !== sourceStat.dev ||
-        targetStat.ino !== sourceStat.ino ||
-        !isInsideRoot(rootPath, canonicalSource)
-      ) {
-        throw new ExternalRootError(
-          "non_verifiable",
-          "The rollback source is not the verified moved object.",
-        );
-      }
-      await this.unlinkIfIdentity(
-        runtime,
-        targetRelativePath,
-        restoredIdentity,
-      );
-    } catch (error) {
-      if (restoredIdentity) {
-        const targetStateAfterFailure = await this.safeMovePathState(
-          runtime,
-          targetRelativePath,
-          snapshot,
-        );
-        if (targetStateAfterFailure === "matching") {
-          await this.unlinkIfIdentity(
-            runtime,
-            sourceRelativePath,
-            restoredIdentity,
-          ).catch(() => undefined);
-        }
-      }
-      if (error instanceof ExternalRootError) throw error;
-      const code =
-        typeof error === "object" && error && "code" in error
-          ? String(error.code)
-          : "";
-      throw new ExternalRootError(
-        code === "EEXIST" ? "target_exists" : "non_verifiable",
-        "The external move rollback could not be completed safely.",
-      );
-    }
+    void snapshot;
+    throw externalMoveMutationUnavailableError();
   }
 
   async recoverMoveToSource(snapshot: ExternalMoveSnapshot): Promise<void> {
-    const runtime = this.requireCapability(snapshot.rootId, "move");
-    const sourceRelativePath = normalizeRelativePath(
-      snapshot.sourceRelativePath,
-    );
-    const targetRelativePath = normalizeRelativePath(
-      snapshot.targetRelativePath,
-    );
-    const source = await this.safeMovePathState(
-      runtime,
-      sourceRelativePath,
-      snapshot,
-    );
-    const target = await this.safeMovePathState(
-      runtime,
-      targetRelativePath,
-      snapshot,
-    );
-    if (source === "matching" && target === "absent") return;
-    if (source === "absent" && target === "matching") {
-      await this.rollbackMove(snapshot);
-      return;
-    }
-    if (source === "matching" && target === "matching") {
-      const sourcePath = await this.resolvePath(runtime, sourceRelativePath);
-      const targetPath = await this.resolvePath(runtime, targetRelativePath);
-      const sourceStat = await lstat(sourcePath, { bigint: true });
-      const targetStat = await lstat(targetPath, { bigint: true });
-      if (
-        sourceStat.isSymbolicLink() ||
-        targetStat.isSymbolicLink() ||
-        sourceStat.dev !== targetStat.dev ||
-        sourceStat.ino !== targetStat.ino
-      ) {
-        throw new ExternalRootError(
-          "non_verifiable",
-          "Source and target do not identify the same recovery object.",
-        );
-      }
-      await this.unlinkIfIdentity(runtime, targetRelativePath, {
-        dev: targetStat.dev,
-        ino: targetStat.ino,
-      });
-      return;
-    }
-    throw new ExternalRootError(
-      "non_verifiable",
-      "The external move cannot be recovered automatically.",
-    );
+    void snapshot;
+    throw externalMoveMutationUnavailableError();
   }
 
   async getPrivateReferenceLocation(
@@ -1276,76 +1065,6 @@ export class ExternalRootsService {
       );
     }
     return targetPath;
-  }
-
-  private async unlinkIfIdentity(
-    runtime: RootRuntime,
-    requestedRelativePath: string,
-    expected: ExternalMoveIdentity,
-  ): Promise<void> {
-    const relativePath = normalizeRelativePath(requestedRelativePath);
-    let absolutePath: string;
-    try {
-      absolutePath = await this.resolvePath(runtime, relativePath);
-    } catch (error) {
-      if (error instanceof ExternalRootError && error.code === "not_found") {
-        return;
-      }
-      throw error;
-    }
-    const current = await lstat(absolutePath, { bigint: true });
-    if (
-      !current.isFile() ||
-      current.isSymbolicLink() ||
-      current.dev !== expected.dev ||
-      current.ino !== expected.ino
-    ) {
-      throw new ExternalRootError(
-        "precondition_failed",
-        "The external path identity changed before cleanup.",
-      );
-    }
-    await unlink(absolutePath);
-  }
-
-  private async safeMovePathState(
-    runtime: RootRuntime,
-    requestedRelativePath: string,
-    snapshot: ExternalMoveSnapshot,
-  ): Promise<SafeMovePathState> {
-    const relativePath = normalizeRelativePath(requestedRelativePath);
-    try {
-      await this.resolveNewTarget(runtime, relativePath);
-      return "absent";
-    } catch (error) {
-      if (
-        !(error instanceof ExternalRootError) ||
-        error.code !== "target_exists"
-      ) {
-        throw error;
-      }
-    }
-
-    let absolutePath: string;
-    try {
-      absolutePath = await this.resolvePath(runtime, relativePath);
-    } catch (error) {
-      if (
-        error instanceof ExternalRootError &&
-        (error.code === "not_a_file" || error.code === "not_found")
-      ) {
-        return error.code === "not_found" ? "absent" : "mismatch";
-      }
-      throw error;
-    }
-    const current = await lstat(absolutePath);
-    if (!current.isFile() || current.isSymbolicLink()) return "mismatch";
-    const verified = await this.readVerifiedBuffer(runtime, relativePath);
-    return verified.buffer.length === snapshot.size &&
-      verified.modifiedAt === snapshot.modifiedAt &&
-      sha256(verified.buffer) === snapshot.sha256
-      ? "matching"
-      : "mismatch";
   }
 
   private async resolvePath(
